+++ tests/Classes/Frontend/Filter/Utility/CategoryTest.php
@@ -124,7 +124,7 @@ class CategoryTest extends BaseTestCase
     /**
      * @return string[][]|number[][]|string[][][]
      */
-    public function dataProviderHandleSysCategoryFilter()
+    public static function dataProviderHandleSysCategoryFilter()
     {
         return [
             [